When we guess the date, make the year consistent
When a user enters a month name to filter by, we
correct this for them, filtering results by the first of
that month. However because of the way the chronic gem calculates
this for us, months prior to today are guessed as 01/01/this year,
and months ahead of today are guessed as 01/01/next year. This
commit corrects that.