Remove user selection from email signup tracking label
The options which a user has selected in the checker are visible in the URL on the results page.
Previously, when clicking one of the email signup links on this page, the full URL was being passed into GA. This was making the GA reports very long and unwieldy as each unique URL is shown as a separate row. In 1 week, over 3000 rows were generated.

PAs have requested that the tracking label is changed so the user selection part of the URL is ommitted. Essentially, we just want to pass in `get-ready-brexit-check/email-signup`.