Add AB test for search popularity
This enables finder-frontend to take an AB test
request header and use it to make requests to
search-api using the AB test variant provided.

This is part of https://docs.publishing.service.gov.uk/manual/run-ab-test.html#2-how-to-set-up-an-ab-test

Once this is deployed the next step will be to
deploy a 100:0 AB test to the CDN.

This is testing a popularity AB test that is
already deployed to production in sesrch-api.

Trello: https://trello.com/c/9BTLARZ8/1064-start-ab-test-for-new-popularity