Restructure SCSS imports
SCSS files located in the root of assets/stylesheets are compiled individually.
In the case of finder_frontend this results in a failure because the imported mixins are not required at the top of the file. Because finder_frontend is not required as a separate stylesheet we include it in the main application.scss as a partial located in views.