Exclude living question choice from travel options
Users have fed back that it is confusing to see the place that they live as an
option when asked where they intend to travel to. This captures the option that
a user selects when asked where they live and removes that location as an
option when they see the travel question. Since this is a presentation issue,
it made sense to construct the options that are displayed in a helper that can
be used by the view.