Rescue 403 errors from content store
At the moment finder-frontend is throwing a 500 when it receives
an error code it doesn't recognise from content-store. That means
that users are shown a "Problem has occurred" page.

We need to handle this error properly so that in the case of a
403, the user will be properly routed to signon and asked to login
if there is any form of access limiting on the content item (e.g.
hosted on draft stack, access limited to organisation)

Note that I've added a test to finders_controller_spec.rb in the
absence of a generic application controller test suite. There are
other places I could have written this test but as they all inherit
from ApplicationController, this would be a lot of unnecessary
duplication.