Add 'exclude_if' on options so we can exclude an option
Adds a very simple optional 'exclude_if' to options in the questions.
This stops an option from showing if the given value has been
previously selected by the user.

This commit uses this functionality to update the questions so that
users are no longer asked if they want to travel to the place
they have indicated they live.