Clarify brexit checker business questions
Changes are intended to do three things:
1) User research has shown that business questions need to be "hardened",
we need to clarify that this only applies to UK businesses, as this has
been a reoccurring point of confusion in user testing.

2) User research shows that Users are often not ticking that they "Sell or
provide services in the UK" even when it should apply to them. If we have
clarified (1), we can assume that this is true and reduce the friction.

3) The descriptive text has proven more confusing than helpful, so we're
removing it