Add A/B test for shingles
This enables finder-frontend to use the shingles AB test request header
to request the shingles variant from search-api when appropriate.

This is part of https://docs.publishing.service.gov.uk/manual/run-ab-test.html#2-how-to-set-up-an-ab-test

Once this is deployed, the next step is to deploy a 100:0 AB test to the CDN.

Trello card: https://trello.com/c/Px5wOXil/1079-run-an-a-b-test-for-shingles.