Remove dynamic filter test
When this test was added, the sidebar didn't automatically reload, and
so people would be unable to select options which had become available
after the initial page load.  The solution to that was to switch off
the dynamic behaviour (0284fd1) and add a test that things weren't
being dynamic (8e1b3c0).

This branch fixes the sidebar reloading issue, so users won't be
incorrectly constrained, meaning this test is now wrong.