Hide keyword tags in A/B test
We want to hide the keyword facet tags from users when they
perform a text search until they've interacted with a filter and have
it selected.

Once they have no more filters selected, we hide the tags again.

This sets up the HideKeywordFacets controller with all the logic and
imports it into the finders_controller.

We attach a CSS class to the groups of facet tags basedcon their
[:data_facet] value, so that we can target keywords group.

We use javascript to check for the A/B test meta tag value. If the
user is in the B variant, we hide the keyword tags in they're the
only tags present (which meant users hasn't used any of the filters)