Separate criteria extraction from validation
We import criteria from strings in a spreadsheet. As such they're untrusted data, that we should validate. However criteria that is embedded in a hash can vary between having symbolic keys and string based keys.

This has led to a bug where a non-existing criteria export-from-eu snuck in to two business results.

This PR both fixes that, and separates criteria extraction from validation so it can be properly unit tested. This also lays some useful groundwork for us to re-use this criteria extraction on the new citizens grouping feature.