Remove default frequency
In https://github.com/alphagov/email-alert-frontend/pull/624 we removed preselection of email frequency in the subscription process. This change meant that the default_frequency param was still accepted, but ignored.

The business finder was the only finder to make use of the default_frequency param - we can now remove this as it doesn't do anything.