Use topic_names_sentence for dynamic filters which have filter values
When signing up to receive updates of research and statistics
we show users that their email subscription is called:
'Statistics with 2 Research and statistics' which looks odd.

This change should make this title:
'Statistics with document types of Statistics (published) and Research'
which looks a little better (still not great).

This is because the research and statistics finder email signup
content item [0] has facet_choices which have filter_values
and is also a dynamic filter type (content_store_document_type).

There's a fair amount of variation in the the email-signup
content items and schema, so it might be nice to do some
refactoring here.

https://trello.com/c/UX2Cd7vr/1211-research-and-stats-subscription-descriptions-are-weird