Fix data-track-action string for bis results
In order to have ecommerce tracking we need a data-track-action attribute that identifies where the action's link appears on the page.
For Individual/Citizen actions we've followed the format "You and your family - Visiting the EU - 1.5 - Guidance".

In this case business actions were being output in the format:
Your business or organisation.2 - Guidance

This format is inconsistent. Instead we should treat business actions as part of a single group, this commit changes the format to:
Your business or organisation - 1.2 - Guidance