Add filter modal HTML markup and styles
When the user decides to filter the results, we show them the available
filters in a modal along with the title and options to return back to
the results or clear all selected filters

In the header we are using CSS Grid to position the title
and return link side-by-side.

In the footer we have bottom-fixed main CTA that shows the dynamic
count of available results.

Some added css complexity with filter toggle content on tablet/desktop
but hopefully to be removed soon.