Add Browse page (OscarSearch)
This is an experimental browse page. It is not ready for
merge.

The intent of this change is to provide two new features
to users:

- a single predicted document as an answer to a query
- a set of search results filtered by topics that we
  think the user is interested in (based on the query)

This work was done by @oscarwyatt in https://github.com/alphagov/collections/pull/858
and has been moved here so we can make some performance
improvements using the features we have added to
finder-frontend for that purpose.

https://trello.com/c/RN2A2b8w/1229-next-steps-for-testing-oscarsearch