Fix discrepancy between stats finder and signup content items
statistics_published is now used by both the finder and
signup content item.

This fixes a minor issue where a user would go from
the finder to the signup page with the filter
content_store_document_type=published_statistics
enabled. This would require us to validate this key
using both the finder content item and the signup
page content item. Now they both use the same key
(statistics_published) it'll be easier to validate
that the key is correct.