Add parser for email alert parameters
Email Alert parameters are a little complex at the moment.

This adds a class to parse and validate user input for the
POST request to create an email subscription list (a
SubscriberList in email-alert-api).

The parser, given user input, returns a hash of keys and
values that have been explicitely permitted either in
the signup page content item.

Because of the diversity of signup page content items
the parser is a little more complex than it should be.

The parser doesn't have the final say on what gets sent
to email-alert-api (that's handled in EmailAlertSignupAPI).
However, it does ensure that the values that do get sent
are allowed.

It also adds a validation that keys and values are
alphanumerics with hyphens. This seems like a good
precaution in case there's a bug in the parser
code or some corrupted data enters the allow list
of a content item.

In future it would be good to refactor this stuff
further.

https://trello.com/c/LMsj6yPk/1247