Use EmailAlertParameterParser for email alert signups
This uses a new parameter parser when a user requests
to fetch or create a subscriber list to subscribe to.

The parser implements an allow list, to ensure that the
user's filters are permitted. If any of the filters
are not explicitely present in the allowlist of the
parser, it will remove them before looking up the
subscriber list.

See commit cb1768a2 for further details about the parser.

https://trello.com/c/LMsj6yPk/1247