Display an error to users if their filters are unprocessable
This will display a generic error to users on email
signup pages should their request to create a subscription
be unprocessable.

This is only likely to happen if they're trying very
strange requests which include unacceptable characters
not present in a finer_email_signup content item allowlist
of filter keys or values.