Add autocomplete suggestions javascript module
A GOV.UK Module to enhance the search box with accessible autocomplete
suggestions from the seach api.

Uses XMLHttpRequest to fetch suggestions from the endpoint and runs on
IE > 10.

Autocomplete suggestions are powered by
https://github.com/alphagov/accessible-autocomplete

We cache suggestions and only fetch from the API when the query has no
matches in the cached array.

We delay the check for matching suggestions by 300ms to minimise the
API calls and/or not update matching results too quickly when users type
relatively quickly.