Don't request finder content item in signup requests
The finder content item is requested when a user signs
up for email alerts.

It's requested because we want to take the default filter
from the finder content item, and re-use it as the default
filter for the email signup content item.

That's not necessary, since we can just use the default
filter field on the signup content item.

This should be merged after https://github.com/alphagov/specialist-publisher/pull/1563
is merged and deployed, since it adds the default
filter to the email signup items.

https://trello.com/c/KOdPubpx/1295-dont-request-finder-content-items-during-email-alert-signup