Update livesearch script form change event
When users are offered suggestions and they use a keyboard
to navigate them we don't wan't to fire a search.
We only want to fire it once they've selected a option.

Enable keyword search on the autocomplete input
if they're not happy with the suggestions or no suggestions are found