Prevent form submit on search button click
As we've replaced the text input with a search component
(input + submit button) we haven't implemented preventing the form submit on button click.

While it hasn't been raised as an issue with users,
we still want to prevent the server-side form submission and use the  existing Ajax fucntionality for fetching results.