Changes rubocop config to disable output safety
Rubocop enforces that we should not use html_safe on the output of a method. However in this case as we entirly control the string that is coming into this from the translation yaml there is no risk of an external person sneaking in some content. This linting rule is stopping us deploying for the 31st jan 2020, so i've killed it