Comment out notifications for actions that have been removed
The checker comes with data integrity checks that say you cannot have a notification for an action that does not exist. However we have temporariliy removed some actions whilst departments update the relevant content ahead of transition. At present, leaving the actions in causes test to fail, blocking deployment. However removing them does not make sense as they'll likely need to be returned tomorrow. This is a half way house, when we do the next import from the spreadsheet a dev should run through each of the notifications and find out if any apply to an action with an ID that was temporaily removed and has been returned