Use facet_connector in email subscription titles
This makes use of a new facet key: facet_connector.
This is intended to be used in cases where 'of' is
not a useful joining word between a facet name and
its values.

For example,
All documents of Statistics
reads better as
All documents filtered by Statistics

This also permits not providing a prefix in a
finder_email_signup content item.

https://trello.com/c/UX2Cd7vr/1211