Update spelling suggestions Enhanced Ecommerce tracking to use page path
At some point in the past we moved our tracking to the path of a page
rather than the ID. This wasn't updated for the search spelling
suggestions, which still used content ID.

This pull request fixes this issue. Unfortunately we do not have access
to anything more than the "base_path" from the content in this position,
though on testing the content in this path has an ID that matches the
content_ids that were being used before.