Reject alpha taxons from the topic selector
In order to open the coronavirus taxon to publishers in Whitehall, we need
to publish it.

The taxon will remain hidden from most things on the public facing site if
it is in the `alpha` phase.  Unfortunately, the topic facet is not one of
those things.

We're OK with it being ingested into the full taxonomy registry because this
is just used to populate titles and other context on faceted search from topics
and not the facets themselves.

The coronavirus taxon will be moved into the live or beta phase when we're happy
that it is up to date with the tagged content.