Reduce content item cache
We did this in expectation of a high load at the time of "The Text Message".

5 minutes is plenty, because content store requests are generally pretty
quick, and we do generally need to be responsive to changes in finder definitions.

We cache these because it's not really necessary for these to be requested
by every search as they don't change _that_ often. Search pages are some
of the heavier weight pages on GOV.UK, and it's good to keep them snappy
if possible.