Retire feeds for the coronavirus topical event
The topical event has been removed, as content is now tagged to the coronavirus
taxonomy topic (and sub-topics) instead. This allows for greater granularity
of notifications, whether that's email or RSS feed.

We've moved email subscriptions over, but we don't want to do that with feeds,
because:
 - the feeds are probably different because publishers' tagging is imperfect
 - we want people to sign up for the feed _they_ want, not the feed we give them
 - we also don't want to build up reams of redirect logic.  I intend to delete this _soon_

We should be able to delete the redirect logic for this feed soon, once all
RSS readers have had a chance to collect it.