Redirect covid topical event searches
The coronavirus topical event has been deprecated for a number of weeks, and
we have now deleted it. We're aware that there may be some remaining links
to search pages using this topical event, and now there are no results appearing
there.

So that we can be good citizens of the web, we should redirect those searches
to the equivalent page using the coronavirus taxon.

This overwrites other taxons that may be present on the search page, but
this approach is better than anything else we can think of.

Other parameters are preserved.