Do not redirect atom feeds
This was not supposed to capture atom feeds, as we had a separate handler
for [retiring the covid topical event feed](https://github.com/alphagov/finder-frontend/pull/2127).
We did this so that people could make an informed choice about which new feed
to choose, rather than just getting the firehose of everything again.

This also pinned the request to the html format, so atom feed requests got an html
response.