Use an h2 for the email subscription header
This is presented as a header (bold, uses the govuk-heading-m class), but uses
a p tag. It seems semantically better to use an h2 as it appears below the page
h1. This also matches the h2 used on the email callout at the bottom of the page.

An alternative would be to not present this as a heading, but this may
require a change to the design, which we don't have time for right now.

There is no visible change.

[Example URL on GOV.UK](https://www.gov.uk/transition-check/results?c%5B%5D=owns-operates-business-organisation&c%5B%5D=visiting-driving&c%5B%5D=visiting-ie&c%5B%5D=visiting-eu&c%5B%5D=visiting-row&c%5B%5D=travel-eu-business-no&c%5B%5D=working-uk&c%5B%5D=living-uk&c%5B%5D=nationality-uk)

https://trello.com/c/EE5x4wxA/444-transition-check-results-heading-looks-like-heading-but-isnt-one