Update to which criteria we display to the user in their grouped actions
- Previously we were creating the list of criteria by combining all of the criteria
for all of the actions displayed in a group. But when one action has multiple grouping, meaning
it is displayed in two groups, the criteria displayed were not always relevant. Eg
Visiting the EU
Because you said:
- You plan to travel to the EU
- You plan to travel to Ireland

Visiting Ireland
Because you said:
- You plan to travel to the EU
- You plan to travel to Ireland

Now, before adding all of the criteria for all of the actions in a group to the list, we filter.
If an action has multiple grouping criteria, we check if the action's criteria shares a key with the
current group. If it does we display the criteria, if it does not we don't.