Add accounts_api to services
We want to be able to have a single place we can go to update the URL
where the accounts application will live at, so that we don't have to go
around replacing all the places we call it in the app.

In many other applications we use Plek for this, but as accounts is new
it's not there yet. So we can use services.

However, we don't want this method to appear unless the feature flag is
set.

This is because we're ultimately going to get this value out of the
environment. If the feature flag isn't set, we wouldn't expect the value
to need to be there. However if the feature flag _is_ set, then must
have the URL of the account manager or else we'll get errors.

It's my preference to use ENV.fetch for this, as it gives you a nice
KeyError in your stack trace and demonstrates clearly to a developer
what they've forgotten to include. The alternative would be an error
message that point we attempt to use Services.accounts_api. In my
experience those stack traces are much harder to follow and take longer
to debug.

So to stop ENV.fetch getting in the way when it shouldn't we're going to
wrap this entire method in a flow control that depends on the feature
flag being picked up in the rails config.