Make heading-like things actual headings
Each action in the results page starts with a line of text / link that
is styled like a heading, but is a `p` tag. The content is designed in a
heading style, and it would be useful for them to be listed as headings
so that they can be recounted by a voice assistant etc, so let's make them
so.