Remove aria-label from Brexit results asides
The labels are not sensible at the moment - they are supposed to be readable
language, not kebab-case slugs.

This also rejigs the markup a little to extract the section heading from the
aside. The heading refers to both the aside, and any actions that are presented
for this group. It makes sense if it is not encapsulated within the aside.