Move Brexit breadcrumbs so skiplinks skips them
I've moved the breadcrumbs/back links on the Brexit checker so that they live
outside the main content section. This means that using skiplinks navigates
you to the start of the content, and not just to a confusing breacrumb link.

I've also updated a link to /get-ready-brexit-check so that it points to
/transition. The old page has been retired, redirecting to the transition landing
page, so this change eliminates that redirected request, while not changing
the outcome.

Describe the problem
--------------------
The main element of the transition check flow includes the back link and the skip link leads to the back link.
The same is happening on the transition check results page but with the breadcrumbs.

Problem category
----------------
Fail of WCAG SC 2.4.1

Impact of the problem
---------------------
Skip link: Keyboard only users cannot skip directly to the main content.
Main element: Screen reader users cannot skip directly to the main content.

https://trello.com/c/meGNv7bk/474-search-results-main-element-and-skip-link-before-main-content
https://trello.com/c/MR7bfTid/475-transition-check-main-element-and-skip-link-before-main-content