Allow different heading levels on Brexit results
Citizen actions need to be h4s because h1, h2 and h3 exist before them in
the page hierarchy.

Unfortunately, because business results do not have groupings, the h3 heading
level does not exist. We can't then use h4s because we'd be [skipping a level](https://www.w3.org/WAI/tutorials/page-structure/headings/#heading-ranks),
which is bad.

The intention is for one of these things to happen:
- redesign business results so that they are grouped in a similar way to citizen results
- add in an appropriate h3 at the top of the business results

This would allow us to use h4s for all actions on the results page, which
I think would be useful for reasons of consistency.

Unfortunately, neither of these things are ready yet.

In the interests of making things better, I propose using h3s and h4s as
an interim measure.