Add service to lookup location data by postcode
The service follows the pattern of other services across GOV.UK for
example [find local council](https://www.gov.uk/find-local-council) and
uses Mapit to do the lookup.

This data will be used to determine if postcode has local Covid-19
restrictions by using the GSS code of the area it is in an mapping that
to the local restrictions data.

Unfortunately, a postcode can be in multiple areas, not all of which
have a GSS code. We considered only returning the area that is
considered a Lower Tier Local Authority (LTLA), but there is no
consistency in how these areas are described in Mapit.
Instead we are returning all areas for a postcode that have a GSS code,
and expect that the caller of this service will loop through all of the
areas and compare each GSS code to the restrictions data.