Add level two restrictions
Adds a yaml file that describes all of the gss codes that are currently under level two restrictions.

I've chosen app/lib as the place for this. This matches what's been done with the brexit checker. We should easily be able to add other levels to this and change things around as needed.