Only set a session cookie when user is logged in
We have enabled cookies as the Rails session store.  This has a side
effect of setting a `_session_id` cookie for all users.  Since this is
not an essential cookie (unless logged in), we need to supress this
cookie until a user is logged in.

Adding to the ApplicationController will mean this cookie will not be
set for any usage, unless a user is specifically logged in.