Implement saving updated results
This commit implements storing new checker results against the
account.

When the user is not logged into the checker:
- If they click the "Create a GOV.UK account" button
  - If they then register:
    - Attributes are set on their account. This PR doesn't touch this
      journey.
  - If they then login:
    - They get redirected back to the checker to confirm the new
      attributes.

- If they click the "sign in if you've saved your answers before"
  link:
  - If they then register:
    - Attributes are not set on their account. They see a link back to
      the checker, which goes to the confirm-new-attributes page.
  - If they then login:
    - They get redirected back to the confirm-new-attributes page.

When the user is logged into the checker:
- If their answers differ from the answers saved in their account:
  - If they click the link to save the new answers:
    - They get sent to the confirm-new-attributes page.

- If their answers match the answers saved in their account:
  - They don't see any call-to-action.

On the confirm-new-attributes page, both the old and new attributes
are shown (currently just as plain text lists) along with a save
button. If the user clicks the button, the OAuth access token in the
session is used to update attribute-store.