Add a :has_session key to the session for tracking login state
We can't drop the session if :sub isn't set, because :sub gets unset
when a user is logged out, which means the updated session never gets
committed and the user is still logged in!

This isn't ideal, but by setting a key when the user logs in (and
never unsetting it), we can ensure that the session gets dropped for
non-accounts users.