Add Subscriber list slug to GOV.UK Account JWT
When a user first registers, we sent them over to the account manager with everything they need to get set up. Prior to this commit that had meant everything we need to save their results if they complete signup and bring them back.

For this feature we also want to offer the user the chance to sign up to email alerts during account creation.

As the checker already has the code to query and get a subscription slug for current email signups, we can use that code to get the slug we'd need to make the subscription and add that to the JWT payload that we send over to the account manager

We also stub refactor email alert API into it's own helper and stub it
out in the tests