Fix test
The previous commit (generated by running the import task), changed the
priorities of some of the actions, including increasing the priority of
T099 which is the subject of this test. We need to update the test to
stay in line with that.

It's useful to be specific about the order here, because it has caught problems
with the ordering logic in the past.