Update email alerts when updating answers
If the user doesn't already have an active subscription, prompt them
to create one.  This isn't quite ideal, as we now have the email
signup frontend and content in both the account manager (for the
registration journey) and here, but we want to rethink this whole
email integration later anyway.

If the user already has an active subscription just update it without
explicitly asking, as we tell them on the confirmation page that this
will happen.