Add post_register_uri to accounts JWT
With this parameter we can get a link back from the post-registration
screen to the transition checker's saved-results page.  This will log
the user into the checker (previously they would only have been logged
into the account manager) and lets us send back additional data in the
query string (though we have to be careful with that because not
everyone will click it).