Manage our own session cookie for accounts
Using the rails session management has a couple of problems: we had to
implement the "has_session" hack to avoid setting the cookie on people
who haven't opted in to accounts; and we can't delete the cookie, only
reset it to a blank session.  It would be nice if the cookie only
existed while a user was logged in, because then we could use "does
the cookie exist?" in the collections app to determine whether to
render the accounts header on /transition.

By managing our own cookies we solve both of those problems: we can
set the cookie on log in and delete it on log out.  We do need to bump
the expiration time ourselves, which is a bit of extra work, but I
think the benefits outweigh that.