Redirect to account manager on sign in if no redirect_path
We have separate session cookies for www.gov.uk and the accounts
domain itself, this is an unfortunate necessity of not being able to
share a single cookie across domains.

By changing the fallback redirect path to the account manager itself,
we can send a user to /transition-check/login to make them log into
both apps:

1. /transition-check/login sends them to the account manager
2. They log in, creating an accounts session cookie
3. Accounts redirects them to /transition-check/callback
4. finder-frontend creates a gov.uk session cookie
5. finder-frontend redirects them back to the account manager

So the user ends up logged into both, and seeing the account manager.
This is what we want the "sign in" link in the GOV.UK header to do, as
only signing into one app would be confusing.