Implement finder-frontend side of single-sign-out
We want signing out of finder-frontend to sign out of the account
manager, and vice versa, for this initial experiment.  This is a bit
unusual for a federated accounts system like this, but for now we are
presenting the GOV.UK account as more of a Transition Checker account.
We will work on the larger design problem of making users understand
that the account is a distinct thing later.

Here's how it'll work.  It works the same on both sides, so I'll just
call the two endpoints "logoutA" and "logoutB".

1. User visits logoutA
2. User is 302-redirected to logoutB?continue=1
3. User's B session cookie is deleted
4. User is redirected to logoutA?done=1
5. User's A session cookie is deleted
6. The post-logout screen or redirect is shown

Redirecting through every signed-in service won't scale beyond 1
service, but for this first experiment, we do only have 1.