Remove broken ecommerce tracking
This is currently causing JS errors when visiting pages like:

- /government/organisations
- /government/groups

and is leading to Smokey test failures, which is masking our
visibility of real issues. Investigation shows this is related
to the size of the request, and removing some parts of the data
appears to fix the issue. Since the data is non-critical, and a
permanent fix is elusive, this removes the tracking attributes
for affected pages to avoid unnecessary errors.