Handle nested options in results comparison
An option can have sub_options.  Currently this only goes one level
deep, but I thought the code would be clearer by solving the general
case of arbitrarily nested options.

This change means that if multiple options are selected, they'll be
sorted by key.  But the key names are similar to the value names, so
that should be ok.