Add cross domain tracking to create account form
Cross domain GA needs IDs visible on the URL when it hands off to
another serivce. This adds some javascript that conditional on the
existance of the account-signup form and GA payloads being present will
go back and overwrite the action / target URL of the create account
form.

The result is an appended _ga parameter with relevant IDs as values