Use name for date filter title
At the moment the title of the date filter is hard coded to the string
"Date". This means that if we want to display two date filters, then we
end up with two filters both with the word "Date".

I'm not really clear on the consequence for this and whether it was
intentional, so it might need some product/frontend approval first.