Update email signup with new design
This is to be consistent with the page for content item signups [1],
when there are no modifiable choices. In particular, while some signup
pages may have a description, we have decided these are too inconsistent
to provide any value, so we have removed them altogether. The design also
calls for the removal of the contextual title, and "alert" wording.

Note that the "govuk-!-font-weight-bold" class is required due to the
unfortunate "font-weight: inherit" CSS reset that pervades this app [2].
It's out-of-scope of this work to explore removing these ancient resets.

[1]: https://github.com/alphagov/email-alert-frontend/pull/918
[2]: https://github.com/alphagov/finder-frontend/commit/ee71a6fce17cba1baf7d5812c7de491039b00adc