Add rake task to generate skeleton notifications
It's a bit of a faff to generate UUIDs and copy/paste configuration in
notifications.yaml every time we need to send an email.

This rake task generates valid notifications yaml that can be copied into
the configuration file and edited rather that having to generate it all
by hand.

Additions should mostly be fine unedited, but changes will need the `note` field
to be amended with the change text that is in the trello card.

If there are custom criteria for a notification (that is, something different
from the criteria set in the corresponding action), then this will still
have to be manually added.

The trello card should have a list of additions/changes that need to be sent,
and these can be passed into the task by setting two environment variables
`NEW_ACTIONS` and `CHANGED_ACTIONS`. Either of these can be missing.

Example usage:

```
rake brexit_checker:configure_notifications NEW_ACTIONS="A001 A099" CHANGED_ACTIONS="S007"
```

You can run `rake -D` to check the full description of the rake task.