Add the cross-domain tracker to potentially cross-domain links
There are some links to www.gov.uk which may redirect to the accounts
domain.  We want cross-domain tracking to work in the case where there
is a redirect, which means we need to pass along the cross-domain _ga
parameter.

Since these are links to www.gov.uk, they don't get decorated by the
normal cross-domain logic.  So instead, instantiate a module from
govuk_publishing_components which does it explicitly.