Fix using subscription_url from email API
This was removed in [1], without updating this app to use the list
slug to form the subscription URL. While it's not ideal to have
shared knowledge of URLs, it's only the case for this app and Email
Alert Frontend, which provides the rest of the signup flow anyway.

Note that there is definitely scope for refactoring here to DRY-up
the repetitious stubbing in the unit and controller specs. However,
the priority for now is to fix the issue we're seeing.

[1]: https://github.com/alphagov/email-alert-api/pull/1517