Some adjustments for the brexit checker templates
The brexit checker templates use the finder_layout layout file.
This layout file already contains a <main> wrapper. However some of the
brexit checker views also contain a main wrapper, which means that on
those pages there would be two main elements, which goes against the
rule that the <main> element must be unique to the document.

In addition to this, these views also contain breadcrumbs. Since
breadcrumbs are a navigation element, they should live outside of
<main>. I updated these views to store the breadcrumbs using content_for.
This way the breadcrumbs will be placed outside of <main> in the layout
file.