Update "title" field with new content
Previously this was used in two places by the email system:

- In the subject line of individual-type emails [1].
- In the body content of the email (prepended to the specified "body") [2].

The new content is the same as the title of all subscriber lists for
the checker, so this change would previously have ended up repeating
the same content in the email body. However, this is no longer the
case since [3]. In future we should consider fully decoupling "title"
and explicitly specifying a "subject" to the email system.

[1]: https://github.com/alphagov/email-alert-api/blob/f55bf44281b282054ff5065f60a93b1819e88e27/app/builders/immediate_email_builder.rb#L36
[2]: https://github.com/alphagov/email-alert-api/blob/f55bf44281b282054ff5065f60a93b1819e88e27/app/presenters/message_presenter.rb#L9
[3]: https://github.com/alphagov/email-alert-api/pull/1538