Run yarn linting as part of default rake task
This moves this repo to the GOV.UK conventional approach of performing
the node linters as part of the rake lint task which is run as part of
rake default.

I also updated the Rakefile to be more conventional by using the modern
Rails.application.load_tasks and clearing the default rake task rather
than adding to it.

I removed the setting to set cucumber as the default task as that will
now have no effect.