Ensure accounts remains enabled for other 5xx errors
In previous commits we've added a check to accounts to remove sign up
options for 503 errors.

503s are specifically temporary and planned, so in the circumstance
where we've pulled the emergency brake on accounts we want the checker
to default gracefully back to just email signups.

However other 5xx errors may be intermittent or unplanned. In these
cases we do not want account signup buttons intermittently disappearing.

We need to think through a better strategy for resilience here in
general (perhaps based on duration / volume of 5xx errors, or perhaps
just relying on monitoring so 2ndline / sentry know when something is
going wrong. But for now, the plan is not to block the user for any 5xx
error OTHER than 503s.