Add requests specs for account suspension
We have implimented an "emergency brake" feature on accounts that will
allow us to suspend the account. This will temporarily remove public
access to accounts to allow us to either perform some maintance or
respond to an incident.

In these cases the account will return a 503 HTTP status code.

The checker should then revert all views back to email only
subscriptions, allowing it to gracefully handle the lack of accounts and
prevent new user journies into the account world.

Routes that are exclusive to checker accounts integration (so for a
small number of users mid-journey when we suspend) we want to halt their
journey and redirect them to the account 503 page which will have a
message explaining what is currently happening with accounts.

The paths targeted here are all those used by the `check_account_enabled`
callback, defining expected behaviour.