Continue to delete sessions if accounts returns 503
In most cases we want to redirect the user to a 503 error page if they
are trying to do "account-y" things and the service is temporarily down,
so they can get a message.

However clearing sessions should be something that can continue. So we
want to make an exception here.